refactor: restructure of project
