+++ src/pages/home/HomePage.tsx
@@ -2,8 +2,8 @@ import { useEffect, useState, type FC } from 'react';
 import { Outlet } from 'react-router';
 import { keepPreviousData, useQuery } from '@tanstack/react-query';
 import { useFilter } from '../../hooks/useFilter';
-import { Pagination } from '../../components/pagination/Pagination';
-import { SearchForm } from '../../components/search/SearchForm/SearchForm';
+import { Pagination } from './components/Pagination/Pagination';
+import { SearchForm } from './components/SearchForm/SearchForm';
 import { Loader } from '../../components/loader/Loader';
 import { ResultsList } from './components/ResultsList/ResultsList';
 import { SelectedFlyout } from './components/SelectedFlyout/SelectedFlyout';
