+++ src/pages/home/components/CharacterDetailsPanel/CharacterDetailsPanel.tsx
