+++ src/pages/home/components/CharacterDetailsPanel/__tests__/CharacterDetailsPanel.test.tsx
