+++ src/pages/home/components/SearchBtn/__tests__/SearchBtn.test.tsx
