+++ src/pages/home/components/SearchField/__tests__/SearchField.test.tsx
